feat: adding typography
