+++ src/docs-ui/app/mat-components/mat-components.component.scss
@@ -1,0 +1,98 @@
+.app-content {
+  padding: 5px;
+}
+
+.app-content mat-card {
+  margin: 5px;
+}
+
+.app-sidenav {
+  min-width: 200px;
+}
+
+.app-content mat-checkbox {
+  margin: 10px;
+}
+
+.app-toolbar-filler {
+  flex: 1 1 auto;
+}
+
+.app-fab {
+  position: fixed !important;
+  bottom: 20px;
+  right: 20px;
+}
+
+.app-spinner {
+  height: 30px;
+  width: 30px;
+  display: inline-block;
+}
+
+.app-input-icon {
+  font-size: 16px;
+  padding: 0 5px;
+}
+
+.app-menu-button {
+  padding: 0;
+  min-width: 50px;
+}
+.app-button {
+  margin: 5px;
+}
+
+.app-list {
+  border: 1px solid rgba(0, 0, 0, 0.12);
+}
+
+.app-progress {
+  margin: 5px;
+}
+
+.example-section {
+  display: flex;
+  align-content: center;
+  align-items: center;
+  height: 60px;
+}
+
+.example-margin {
+  margin: 10px;
+}
+
+.mat-slider-horizontal {
+  width: 300px;
+}
+
+.mat-slider-vertical {
+  height: 300px;
+}
+
+mat-chip {
+  max-width: 200px;
+}
+
+.app-sidebar-button {
+  width: inherit;
+}
+
+.example-spacer {
+  flex: 1 1 auto;
+}
+
+.example-h4 {
+  letter-spacing: 1px;
+  font-weight: 200;
+  font-size: 14px;
+}
+
+.example-card {
+  max-width: 400px;
+}
+
+.example-header-image {
+  background-image: url('https://material.angular.io/assets/img/examples/shiba1.jpg');
+  background-size: cover;
+}
