+++ src/docs-ui/app/typography/typography.component.html
@@ -1,0 +1,21 @@
+<div layout="row" class="pad-top">
+  <div class="pad-md push">
+    <div class="mat-display-4">Display-4</div>
+    <div class="mat-display-3">Display-3</div>
+    <div class="mat-display-2">Display-2</div>
+    <div class="mat-display-1">Display-1</div>
+    <div class="mat-headline">Headline</div>
+    <div class="mat-title">Title</div>
+  </div>
+  <div class="pad-md push">
+    <div class="mat-subheading-2">Subheading-2</div>
+    <div class="mat-subheading-1">Subheading-1</div>
+    <p class="mat-body-2">Body-2</p>
+    <p class="mat-body-1">Body-1</p>
+    <mat-form-field>
+      <input matInput placeholder="Input" />
+    </mat-form-field>
+    &nbsp;
+    <button mat-raised-button>Button</button>
+  </div>
+</div>
