+++ src/docs-ui/app/typography/typography.component.scss
@@ -1,0 +1,13 @@
+.types {
+  display: grid;
+  grid-template-areas: 'left right';
+  grid-template-columns: 1fr 1fr;
+
+  .headers {
+    grid-area: left;
+  }
+
+  .bodies {
+    grid-area: right;
+  }
+}
