+++ src/docs-ui/app/typography/typography.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-typography',
+  templateUrl: './typography.component.html',
+  styleUrls: ['./typography.component.scss'],
+})
+export class TypographyComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit() {}
+}
